requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dlib==19.24.5
face-recognition==1.3.0
face_recognition_models @ file:///C:/Users/June%20Nguyen/Downloads/face_recognition_models-master
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.1
opencv-python==4.10.0.84
pillow==10.4.0
setuptools==72.1.0
Werkzeug==3.0.3
